Refactor API classes into a separate file
