+++ lib/src/analysis_server.dart
@@ -5,14 +5,15 @@
 /// A wrapper around an analysis server instance
 library services.analysis_server;
 
-import 'dart:io' as io;
 import 'dart:async';
 import 'dart:collection';
 import 'dart:convert';
-import 'package:logging/logging.dart';
-import 'package:rpc/rpc.dart';
+import 'dart:io' as io;
 
 import 'package:analysis_server/src/protocol.dart';
+import 'package:logging/logging.dart';
+
+import 'api_classes.dart';
 
 /**
  * Type of callbacks used to process notifications.
@@ -497,111 +498,6 @@ class _Server {
     }
     _recordedStdio.add(line);
   }
-}
-
-class CompleteResponse {
-  @ApiProperty(description: 'The offset of the start of the text to be replaced.')
-  final int replacementOffset;
-
-  @ApiProperty(description: 'The length of the text to be replaced.')
-  final int replacementLength;
-
-  final List<Map<String, String>> completions;
-
-  CompleteResponse(this.replacementOffset, this.replacementLength,
-      List<Map> completions) :
-    this.completions = _convert(completions);
-
-  /**
-   * Convert any non-string values from the contained maps.
-   */
-  static List<Map<String, String>> _convert(List<Map> list) {
-    return list.map((m) {
-      Map newMap = {};
-      for (String key in m.keys) {
-        var data = m[key];
-        // TODO: Properly support Lists, Maps (this is a hack).
-        if (data is Map || data is List) {
-          data = JSON.encode(data);
-        }
-        newMap[key] = '${data}';
-      }
-      return newMap;
-    }).toList();
-  }
-}
-
-class FixesResponse {
-  final List<ProblemFix> fixes;
-
-  FixesResponse(List<AnalysisErrorFixes> analysisErrorFixes) :
-    this.fixes = _convert(analysisErrorFixes);
-
-  /**
-   * Convert any non-string values from the contained maps.
-   */
-  static List<ProblemFix> _convert(List<AnalysisErrorFixes> list) {
-    var problemsAndFixes = new List<ProblemFix>();
-    list.forEach((analysisErrorFix)
-        => problemsAndFixes.add(_convertAnalysisErrorFix(analysisErrorFix)));
-
-    return problemsAndFixes;
-  }
-
-  static ProblemFix _convertAnalysisErrorFix(AnalysisErrorFixes analysisErrFix) {
-    String problemMessage = analysisErrFix.error.message;
-    int problemOffset = analysisErrFix.error.location.offset;
-    int problemLength = analysisErrFix.error.location.length;
-
-    List<Fix> possibleFixes = new List<Fix>();
-
-    for (var sourceChange in analysisErrFix.fixes) {
-      List<Edit> edits = new List<Edit>();
-      for (var sourceFileEdit in sourceChange.edits) {
-        for (var sourceEdit in sourceFileEdit.edits) {
-          edits.add(new Edit(
-              sourceEdit.offset,
-              sourceEdit.length,
-              sourceEdit.replacement));
-        }
-      }
-      Fix possibleFix = new Fix(sourceChange.message, edits);
-      possibleFixes.add(possibleFix);
-    }
-
-    return new ProblemFix(
-        possibleFixes,
-        problemMessage,
-        problemOffset,
-        problemLength);
-  }
-
-}
-
-/// Set of fixes for a particular problem.
-class ProblemFix {
-  final List<Fix> fixes;
-  final String problemMessage;
-  final int offset;
-  final int length;
-
-  ProblemFix(this.fixes, this.problemMessage, this.offset, this.length);
-}
-
-/// A single way of fixing a particular problem
-class Fix {
-  final String message;
-  final List<Edit> edits;
-
-  Fix(this.message, this.edits);
-}
-
-class Edit {
-  final int offset;
-  final int length;
-  final String replacement;
-
-  Edit(this.offset, this.length, this.replacement);
 }
