+++ test/analyzer_test.dart
@@ -9,6 +9,8 @@ import 'package:services/src/common.dart';
 import 'package:cli_util/cli_util.dart' as cli_util;
 import 'package:unittest/unittest.dart';
 
+import 'package:services/src/api_classes.dart';
+
 String sdkPath = cli_util.getSdkDir([]).path;
 
 void defineTests() {
